CA-118 split hitinfo and intersects
