+++ include/intersection/hitinfo.h
@@ -13,10 +13,7 @@ struct HitInfo {
 
     /** @brief Given a face's outward normal and the initial ray, sets front_face to represent
     if collision hits it from the front or not. */
-    inline void set_face_normal(const ray& r, const vec3& outward_normal) {
-        front_face = dot(r.direction(), outward_normal) < 0;
-        normal = front_face ? outward_normal : -outward_normal;
-    }
+    inline void set_face_normal(const ray& r, const vec3& outward_normal);
 };
