+++ src/intersection/hitinfo.cc
@@ -1,0 +1,6 @@
+#include "hitinfo.h"
+
+void HitInfo::set_face_normal(const ray& r, const vec3& outward_normal) {
+    front_face = dot(r.direction(), outward_normal) < 0;
+    normal = front_face ? outward_normal : -outward_normal;
+}
